+++ test/main_tb/tb_top_simple_ctu_can_fd.vhd
@@ -257,7 +257,7 @@ begin
         res_n       => res_n,
 
         -- DFT support
-        scan_enable => scan_enable,
+        scan_mode   => scan_enable,
 
         -- Memory interface
         data_in     => write_data,